renaming file matching conventions
