+++ lib/strava/web/raise_response_error.rb
